// src/modules/transactions/dto/enum.dto.ts
export enum TransactionStatus {
  PENDING = 'PENDING',
  SUCCESS = 'SUCCESS',
  FAILED = 'FAILED'
}

export enum Coin {
  DEMO = 'DEMO',
  SOL = 'SOL',
  SOLR = 'SOLR'
}

export enum TokenAddress {
  SOL = 'So11111111111111111111111111111111111111112',
  SOLR = ''
}

export enum TransactionType {
  INIT_BALANCE = 'INIT_BALANCE',
  PLACE_ORDER = 'PLACE_ORDER',
  PROFIT_ORDER = 'PROFIT_ORDER',
  SYSTEM = 'SYSTEM',
  DIRECT_F1_COMMISSION = 'DIRECT_F1_COMMISSION',
  DIRECT_F2_COMMISSION = 'DIRECT_F2_COMMISSION',
  DIRECT_F3_COMMISSION = 'DIRECT_F3_COMMISSION',
  BINARY_COMMISSION = 'BINARY_COMMISSION',
  RANK_COMMISSION = 'RANK_COMMISSION',
  PEER_COMMISSION = 'PEER_COMMISSION',
  WITHDRAW = 'WITHDRAW',
  WITHDRAW_PROFIT = 'WITHDRAW_PROFIT',
  FIXED_INTEREST_REWARD = 'FIXED_INTEREST_REWARD',
  MULTIPLIER_REWARD = 'MULTIPLIER_REWARD',
  TOKEN_REWARD = 'TOKEN_REWARD',
  WITHDRAW_FEE = 'WITHDRAW_FEE'
}

export enum FundOperation {
  ADD = 'ADD',
  SUB = 'SUB'
}

export enum ERmqQueueName {
  TRANSACTION = 'solr.transaction.queue',
  TRANSFER = 'solr.transfer.queue',
  TRANSFER_FEE = 'solr.transfer-fee.queue'
}

export enum ERmqPrefetch {
  CHAT_GATEWAY = 500,
  NOTIFICATION = 500,
  TRANSACTION = 500,
  TRANSFER = 10,
  TRANSFER_FEE = 10
}

export enum COMMISSION_TYPES {
  DIRECT_F1_COMMISSION = 'DIRECT_F1_COMMISSION',
  DIRECT_F2_COMMISSION = 'DIRECT_F2_COMMISSION',
  DIRECT_F3_COMMISSION = 'DIRECT_F3_COMMISSION',
  BINARY_COMMISSION = 'BINARY_COMMISSION',
  RANK_COMMISSION = 'RANK_COMMISSION',
  PEER_COMMISSION = 'PEER_COMMISSION'
}

export enum WithdrawType {
  PROFIT = 'PROFIT',
  COMMISSION = 'COMMISSION'
}